authentication added in server controller, user controller

+++ server/controllers/server_controller.js
@@ -7,12 +7,12 @@ export const home = function (req, res) {
 // working
 export const createServer = async function (req, res) {
   try {
-    let creator = await User.findById(req.body.user_id);
+    let creator = await User.findOne({ email: req.body.user_email });
     // console.log("reached 1");
     let server = await Server.create({
       name: req.body.server_name,
-      members: [req.body.user_id], // user id of the creator
-      admin: [req.body.user_id], // user id of the creator
+      members: [creator._id], // user id of the creator
+      admin: [creator._id], // user id of the creator
     });
     // console.log("reached 2");
     await User.updateOne(
@@ -24,23 +24,45 @@ export const createServer = async function (req, res) {
       }
     );
     // add this new server to creator's server list
-    res.status(200);
-    res.send("server created");
+    return res.status(201).send({
+      success: true,
+      message: "Server successfully created",
+      server: server,
+    });
   } catch (err) {
-    res.status(404);
-    res.send(`error aara vai : ${err}`);
+    return res.status(404).send({
+      success: false,
+      message: `Bhai error aara : ${err}`,
+    });
   }
   // res.send("Server is Working");
 };
 
 export const deleteServer = async function (req, res) {
   try {
+    // check whther admin or not
+    let server = await Server.findById(req.body.server_id);
+    let present = await Server.find({
+      _id: req.body.server_id,
+      admin: { $in: [req.body.user_id] },
+    }).count();
+    if (present == 0) {
+      return res.status(401).send({
+        success: false,
+        message: "Only admins can delete a server",
+      });
+    }
+    // delete
     await Server.deleteOne({ _id: req.body.server_id });
-    res.status(200);
-    res.send("server deleted");
+    return res.status(201).send({
+      success: true,
+      message: "Server successfully deleted",
+    });
   } catch (err) {
-    res.status(404);
-    res.send(`error aara vai : ${err}`);
+    return res.status(404).send({
+      success: false,
+      message: `Bhai error aara : ${err}`,
+    });
   }
 };
 // not tested
@@ -73,6 +95,8 @@ export const removeAdmin = async function (req, res) {
 // working
 export const addMember = async function (req, res) {
   try {
+    let server = await Server.findById(req.body.server_id);
+    // add member
     await Server.updateOne(
       {
         _id: req.body.server_id,
@@ -89,11 +113,16 @@ export const addMember = async function (req, res) {
         $push: { servers: req.body.server_id },
       }
     );
-    res.status(200);
-    res.send("member added");
+    return res.status(201).send({
+      success: true,
+      message: `You are added into the server ${server.name} Successfully`,
+      server: server,
+    });
   } catch (err) {
-    res.status(404);
-    res.send(`error aara vai : ${err}`);
+    return res.status(401).send({
+      success: false,
+      message: `Bhai error aara : ${err}`,
+    });
   }
 };
